+++ frame_forge/__init__.py
@@ -48,6 +48,7 @@ class GenerateImages:
         sub_strikeout: int,
         sub_scale_x: int,
         sub_scale_y: int,
+        sub_spacing: int,
         source_sub_title: str,
         release_sub_title: str,
     ):
@@ -87,6 +88,7 @@ class GenerateImages:
         self.sub_strikeout = sub_strikeout
         self.sub_scale_x = sub_scale_x
         self.sub_scale_y = sub_scale_y
+        self.sub_spacing = sub_spacing
         self.source_sub_title = source_sub_title
         self.release_sub_title = release_sub_title
 
@@ -132,7 +134,7 @@ class GenerateImages:
         selected_sub_style = (
             f"{self.sub_font_name},{self.sub_size},{color},{secondary_color},{outline_color},{back_color},"
             f"{self.sub_bold},{self.sub_italic},{self.sub_underline},{self.sub_strikeout},"
-            f"{self.sub_scale_x},{self.sub_scale_y},0,0,1,1,0,{self.sub_alignment},10,10,10,1"
+            f"{self.sub_scale_x},{self.sub_scale_y},{self.sub_spacing},0,1,1,0,{self.sub_alignment},10,10,10,1"
         )
         sync_sub_base = (
             "Segoe UI,{size},&H31FF31&,&H00000000,&H00000000,&H00000000,"